# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Reforest API",
    "description": "API for managing tree adoptions, forest plots, and sponsor transactions for the Reforest project.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          }
        }
      },
      "RegisterRequest": {
        "type": "object",
        "properties": {
          "email": { "type": "string", "format": "email" },
          "password": { "type": "string", "minLength": 6 },
          "role": { "type": "string", "enum": ["ADMIN", "SPONSOR"] },
          "full_name": { "type": "string" },
          "date_of_birth": {
            "type": "string",
            "format": "date",
            "example": "2000-01-30"
          }
        },
        "required": ["email", "password", "role"]
      },
      "RegisterResponse": {
        "type": "object",
        "properties": {
          "message": { "type": "string" },
          "user_id": { "type": "string" }
        }
      },
      "LoginRequest": {
        "type": "object",
        "properties": {
          "email": { "type": "string", "format": "email" },
          "password": { "type": "string" }
        },
        "required": ["email", "password"]
      },
      "AuthResponse": {
        "type": "object",
        "properties": {
          "token": { "type": "string" },
          "role": { "type": "string" }
        }
      },
      "Species": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "readOnly": true },
          "common_name": { "type": "string" },
          "space_required_m2": { "type": "number", "format": "float" },
          "price": { "type": "integer" }
        }
      },
...
      "Plot": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "readOnly": true },
          "location_name": { "type": "string" },
          "address": { "type": "string" },
          "available_space_m2": { "type": "number", "format": "float" }
        }
      },
...
      "Tree": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "sponsor_id": { "type": "string" },
          "species_id": { "type": "string" },
          "plot_id": { "type": "string" },
          "custom_name": { "type": "string" },
          "current_height_meters": { "type": "number", "format": "float" },
          "last_care_date": { "type": "string", "format": "date-time" },
          "adopted_at": { "type": "string", "format": "date-time" }
        }
      },
      "TreeList": {
        "type": "object",
        "properties": {
          "trees": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/Tree" }
          }
        }
      },
...
      "CreateLogRequest": {
        "type": "object",
        "properties": {
          "adopted_tree_id": {
            "type": "string",
            "description": "This is ignored, the ID from the path is used."
          },
          "current_height_meters": { "type": "number", "format": "float" },
          "activity": { "type": "string" },
          "note": { "type": "string" }
        }
      },
      "LogList": {
        "type": "object",
        "properties": {
          "logs": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/LogEntry" }
          }
        }
      },
      "AdoptTreeRequest": {
        "type": "object",
        "properties": {
          "species_id": { "type": "string" },
          "plot_id": { "type": "string" },
          "custom_name": { "type": "string" }
        }
      },
      "AdoptTreeResponse": {
        "type": "object",
        "properties": {
          "tree_id": { "type": "string" },
          "payment_url": { "type": "string" },
          "invoice_id": { "type": "string" }
        }
      },
      "BalanceResponse": {
        "type": "object",
        "properties": {
          "balance": { "type": "number" }
        }
      },
      "TopUpRequest": {
        "type": "object",
        "properties": {
          "amount": { "type": "integer" },
          "duration_seconds": { "type": "integer" }
        }
      },
      "Transaction": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "user_id": { "type": "string" },
          "amount": { "type": "integer" },
          "type": { "type": "string" },
          "status": { "type": "string" },
          "payment_url": { "type": "string" },
          "expires_at": { "type": "string", "format": "date-time" },
          "reference_id": { "type": "string" },
          "invoice_id": { "type": "string" }
        }
      },
      "TransactionList": {
        "type": "object",
        "properties": {
          "transactions": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/Transaction" }
          }
        }
      }
    }
  },
...
  "paths": {
    "/auth/register": {
      "post": {
        "summary": "User Registration",
        "tags": ["Auth"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/RegisterRequest" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User created successfully.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/RegisterResponse" }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          }
        }
      }
    },
    "/auth/login": {
      "post": {
        "summary": "User Login",
        "tags": ["Auth"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/LoginRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Returns JWT token and user role.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/AuthResponse" }
              }
            }
          },
          "401": {
            "description": "Invalid credentials",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          }
        }
      }
    },
...
    "/species/{id}": {
      "get": {
        "summary": "Get species details by ID",
        "tags": ["Public"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Detailed information about the species.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Species" }
              }
            }
          },
          "404": { "description": "Species not found." }
        }
      }
    },
...
    "/plots/{id}": {
      "get": {
        "summary": "Get plot details by ID",
        "tags": ["Public"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Detailed information about the plot.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Plot" }
              }
            }
          },
          "404": { "description": "Plot not found." }
        }
      }
    },
    "/trees": {
      "get": {
        "summary": "List all adopted trees",
        "tags": ["Public"],
        "responses": {
          "200": {
            "description": "List of all trees.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/TreeList" }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Adopt a new tree",
        "security": [{ "bearerAuth": [] }],
        "tags": ["Sponsor"],
        "description": "Sponsor action. Creates an adoption intent and returns a payment response. The user's role must be 'SPONSOR'.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/AdoptTreeRequest" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Adoption intent created, payment required.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/AdoptTreeResponse" }
              }
            }
          },
          "401": { "description": "Unauthenticated" },
          "403": { "description": "Permission Denied (not a SPONSOR)" }
        }
      }
    },
    "/trees/{id}": {
      "get": {
        "summary": "Get tree details",
        "tags": ["Public"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Tree details.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Tree" }
              }
            }
          },
          "404": { "description": "Tree not found." }
        }
      }
    },
    "/trees/{id}/logs": {
      "get": {
        "summary": "Get tree maintenance/growth logs",
        "tags": ["Public"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "List of growth and care logs.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/LogList" }
              }
            }
          }
        }
      }
    },
    "/wallet/balance": {
      "get": {
        "summary": "Get user's current wallet balance",
        "security": [{ "bearerAuth": [] }],
        "tags": ["Sponsor"],
        "responses": {
          "200": {
            "description": "Current wallet balance.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/BalanceResponse" }
              }
            }
          },
          "401": { "description": "Unauthenticated" }
        }
      }
    },
    "/wallet/topup": {
      "post": {
        "summary": "Top up wallet balance",
        "security": [{ "bearerAuth": [] }],
        "tags": ["Sponsor"],
        "description": "Generates a Xendit invoice to top up the user's wallet.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/TopUpRequest" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Invoice created. Returns a transaction object with a Xendit payment URL.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Transaction" }
              }
            }
          },
          "401": { "description": "Unauthenticated" }
        }
      }
    },
    "/wallet/transactions": {
      "get": {
        "summary": "View transaction history",
        "security": [{ "bearerAuth": [] }],
        "tags": ["Sponsor"],
        "responses": {
          "200": {
            "description": "List of all financial movements for the user.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/TransactionList" }
              }
            }
          },
          "401": { "description": "Unauthenticated" }
        }
      }
    },
    "/webhooks/finance": {
      "post": {
        "summary": "Xendit Payment Callback",
        "tags": ["Webhooks"],
        "description": "Webhook endpoint for Xendit to post invoice status updates (e.g., PAID, EXPIRED).",
        "requestBody": {
          "description": "Xendit invoice callback payload",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "external_id": { "type": "string" },
                  "status": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Webhook received.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": { "type": "string" }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/admin/species": {
      "post": {
        "summary": "Create a new tree species",
        "security": [{ "bearerAuth": [] }],
        "tags": ["Admin"],
        "description": "Admin action. User role must be 'ADMIN'.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/Species" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Species created.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Species" }
              }
            }
          },
          "401": { "description": "Unauthenticated" },
          "403": { "description": "Permission Denied" }
        }
      }
    },
    "/admin/species/{id}": {
      "put": {
        "summary": "Update a tree species",
        "security": [{ "bearerAuth": [] }],
        "tags": ["Admin"],
        "description": "Admin action. User role must be 'ADMIN'.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/Species" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Species updated successfully.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Species" }
              }
            }
          },
          "404": { "description": "Species not found." }
        }
      },
      "delete": {
        "summary": "Delete a tree species",
        "security": [{ "bearerAuth": [] }],
        "tags": ["Admin"],
        "description": "Admin action. User role must be 'ADMIN'.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "204": { "description": "Species deleted successfully." },
          "404": { "description": "Species not found." }
        }
      }
    },
    "/admin/plots": {
      "post": {
        "summary": "Create a new forest plot",
        "security": [{ "bearerAuth": [] }],
        "tags": ["Admin"],
        "description": "Admin action. User role must be 'ADMIN'.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/Plot" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Plot created.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Plot" }
              }
            }
          }
        }
      }
    },
    "/admin/plots/{id}": {
      "put": {
        "summary": "Update a forest plot",
        "security": [{ "bearerAuth": [] }],
        "tags": ["Admin"],
        "description": "Admin action. User role must be 'ADMIN'.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/Plot" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Plot updated successfully.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Plot" }
              }
            }
          },
          "404": { "description": "Plot not found." }
        }
      },
      "delete": {
        "summary": "Delete a forest plot",
        "security": [{ "bearerAuth": [] }],
        "tags": ["Admin"],
        "description": "Admin action. User role must be 'ADMIN'.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "204": { "description": "Plot deleted successfully." },
          "404": { "description": "Plot not found." }
        }
      }
    },
    "/admin/trees/{id}": {
      "put": {
        "summary": "Update tree details (Admin)",
        "security": [{ "bearerAuth": [] }],
        "tags": ["Admin"],
        "description": "Admin action. User role must be 'ADMIN'.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/Tree" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Tree updated successfully.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Tree" }
              }
            }
          },
          "404": { "description": "Tree not found." }
        }
      },
      "delete": {
        "summary": "Delete a tree (Admin)",
        "security": [{ "bearerAuth": [] }],
        "tags": ["Admin"],
        "description": "Admin action. User role must be 'ADMIN'.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "204": { "description": "Tree deleted successfully." },
          "404": { "description": "Tree not found." }
        }
      }
    },
    "/admin/trees/{id}/logs": {
      "post": {
        "summary": "Add a log entry for a tree (Admin)",
        "security": [{ "bearerAuth": [] }],
        "tags": ["Admin"],
        "description": "Admin action. User role must be 'ADMIN'.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateLogRequest" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Log entry created.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/LogEntry" }
              }
            }
          }
        }
      }
    },
    "/admin/logs/{id}": {
      "put": {
        "summary": "Update a log entry (Admin)",
        "security": [{ "bearerAuth": [] }],
        "tags": ["Admin"],
        "description": "Admin action. User role must be 'ADMIN'.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/LogEntry" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Log updated successfully.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/LogEntry" }
              }
            }
          },
          "404": { "description": "Log not found." }
        }
      },
      "delete": {
        "summary": "Delete a log entry (Admin)",
        "security": [{ "bearerAuth": [] }],
        "tags": ["Admin"],
        "description": "Admin action. User role must be 'ADMIN'.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "204": { "description": "Log deleted successfully." },
          "404": { "description": "Log not found." }
        }
      }
    }
  }
}